+++ src/components/main/payment/StripeSettings.js
